Estimate IV for position width and add limit order to rebalance if necessary (#1)

+++ contracts/libraries/LiquidityAmounts.sol
@@ -83,15 +83,17 @@ library LiquidityAmounts {
         uint160 sqrtRatioAX96,
         uint160 sqrtRatioBX96,
         uint128 liquidity
-    ) internal pure returns (uint256 amount0) {
+    ) internal pure returns (uint224 amount0) {
         if (sqrtRatioAX96 > sqrtRatioBX96) (sqrtRatioAX96, sqrtRatioBX96) = (sqrtRatioBX96, sqrtRatioAX96);
 
         return
-            FullMath.mulDiv(
-                uint256(liquidity) << FixedPoint96.RESOLUTION,
-                sqrtRatioBX96 - sqrtRatioAX96,
-                sqrtRatioBX96
-            ) / sqrtRatioAX96;
+            uint64(
+                FullMath.mulDiv(
+                    uint256(liquidity) << FixedPoint96.RESOLUTION,
+                    sqrtRatioBX96 - sqrtRatioAX96,
+                    sqrtRatioBX96
+                ) / sqrtRatioAX96
+            );
     }
 
     /// @notice Computes the amount of token1 for a given amount of liquidity and a price range
@@ -103,10 +105,10 @@ library LiquidityAmounts {
         uint160 sqrtRatioAX96,
         uint160 sqrtRatioBX96,
         uint128 liquidity
-    ) internal pure returns (uint256 amount1) {
+    ) internal pure returns (uint192 amount1) {
         if (sqrtRatioAX96 > sqrtRatioBX96) (sqrtRatioAX96, sqrtRatioBX96) = (sqrtRatioBX96, sqrtRatioAX96);
 
-        return FullMath.mulDiv(liquidity, sqrtRatioBX96 - sqrtRatioAX96, FixedPoint96.Q96);
+        return uint192(FullMath.mulDiv(liquidity, sqrtRatioBX96 - sqrtRatioAX96, FixedPoint96.Q96));
     }
 
     /// @notice Computes the token0 and token1 value for a given amount of liquidity, the current
